GTFS commands description edited

+++ app/Console/Commands/DownloadGtfsCommand.php
@@ -34,15 +34,30 @@ class DownloadGtfsCommand extends Command
         $api = json_decode(file_get_contents('https://transport.data.gouv.fr/api/datasets/616d6116452cadd5c04b49b7'), true);
 
         if ($hashCode == 61 && $file == 0 && (isset($api) && gettype($api) == "array")){
-            $copy = Storage::copy('gtfs/zip/gtfs-smtc.zip', 'gtfs/zip/'.date('YmdHis').'-gtfs-smtc.zip');
-            if ($copy){
-                file_put_contents(Storage::path('gtfs/zip/gtfs-smtc.zip'), file_get_contents($api['resources'][0]['original_url']));
-                Log::channel('gtfs')->info('File downloaded successfully');
-                $this->line('<fg=green;options=bold>File downloaded successfully');
-            } else {
-                Log::channel('gtfs')->info('File downloading failed');
-                $this->line('<fg=red>File downloading failed');
+
+            $confirm = $this->confirm('Do you want to proceed to an update ?');
+
+            if ($confirm){
+
+                $newName = 'gtfs/zip/'.date('YmdHis').'-gtfs-smtc.zip';
+                $copy = Storage::copy('gtfs/zip/gtfs-smtc.zip', $newName);
+
+                if ($copy){
+
+                    $this->line('<fg=green>The main GTFS archive has been copied as ' . $newName);
+                    Log::channel('gtfs')->info('GTFS archive downloaded successfully as gtfs-smtc.zip');
+
+                    file_put_contents(Storage::path('gtfs/zip/gtfs-smtc.zip'), file_get_contents($api['resources'][0]['original_url']));
+
+                    Log::channel('gtfs')->info('GTFS archive downloaded successfully as gtfs-smtc.zip');
+                    $this->line('<fg=green;options=bold>GTFS archive downloaded successfully as gtfs-smtc.zip');
+
+                } else {
+                    Log::channel('gtfs')->info('GTFS archive download failed');
+                    $this->line('<fg=red>GTFS archive download failed');
+                }
             }
+
         } elseif($hashCode == 60) {
             Log::channel('gtfs')->info('No download required');
             $this->line("<fg=green>No download required");
